primary_jah_agent_core_implementation.py
#!/usr/bin/env python3
"""
Primary JAH Agent - Central Coordination Hub
=============================================

The Primary JAH Agent serves as the CEO-level decision maker and central coordinator
for the autonomous JAH Agency system. It manages all specialized sub-agents, 
orchestrates tasks, and maintains system-wide operational oversight.

Author: JAHA Development Team
Version: 2.0.0
License: MIT
"""

import asyncio
import json
import logging
import time
from datetime import datetime, timedelta
from typing import Dict, List, Optional, Any
from dataclasses import dataclass, asdict
from enum import Enum
import yaml
import uuid

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    handlers=[
        logging.FileHandler('jah_agency.log'),
        logging.StreamHandler()
    ]
)
logger = logging.getLogger(__name__)


class AgentStatus(Enum):
    """Agent operational status enumeration"""
    IDLE = "idle"
    ACTIVE = "active"
    BUSY = "busy"
    MAINTENANCE = "maintenance"
    ERROR = "error"


class TaskPriority(Enum):
    """Task priority levels"""
    CRITICAL = 1
    HIGH = 2
    MEDIUM = 3
    LOW = 4


@dataclass
class SystemMetrics:
    """System performance and operational metrics"""
    active_agents: int
    daily_revenue: float
    completed_tasks: int
    pending_tasks: int
    system_uptime: float
    cpu_usage: float
    memory_usage: float
    error_count: int
    success_rate: float
    timestamp: datetime


@dataclass
class Task:
    """Task representation with metadata"""
    task_id: str
    task_type: str
    priority: TaskPriority
    assigned_agent: Optional[str]
    status: str
    created_at: datetime
    deadline: Optional[datetime]
    payload: Dict[str, Any]
    estimated_duration: Optional[int]
    actual_duration: Optional[int]


@dataclass
class AgentConfig:
    """Agent configuration structure"""
    agent_id: str
    agent_type: str
    specialization: str
    resource_allocation: float
    max_concurrent_tasks: int
    capabilities: List[str]
    status: AgentStatus
    performance_metrics: Dict